compat tidyup
